online_hackathon_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
cx_Oracle==8.3.0
python-dotenv==1.0.0
# 백엔드